+++ src/fst/slice_stmtish.py
@@ -958,7 +958,7 @@ def _elif_to_else_if(self: fst.FST, docstr: bool | Literal['strict'] = True) -> 
     self._put_src([indent + 'else:', indent + self.root.indent], ln, 0, ln, col, False)
 
 
-def _normalize_block(self: fst.FST, field: str = 'body', *, indent: str | None = None) -> None:
+def _normalize_block(self: fst.FST, field: str = 'body', *, indent: str | None = None) -> bool:
     """Move statements on the same logical line as a block open to their own line, e.g:
     ```
     if a: call()
@@ -970,19 +970,23 @@ def _normalize_block(self: fst.FST, field: str = 'body', *, indent: str | None =
     ```
 
     **Parameters:**
-    - `field`: Which block to normalize (`'body'`, `'orelse'`, `'handlers'`, `'finalbody'`).
+    - `field`: Which block to normalize (`'body'`, `'orelse'`, `'finalbody'`). `'handlers'` or `'cases'` wouldn't make
+        sense.
     - `indent`: The indentation to use for the relocated line if already known, saves a call to `_get_block_indent()`.
+
+    **Returns:**
+    - `bool`: Whether block was normalized or not.
     """
 
     if self.a.__class__ in ASTS_LEAF_MOD or not (block := getattr(self.a, field)) or not isinstance(block, list):
-        return
+        return False
 
     b0 = block[0].f
     b0_ln, b0_col, _, _ = b0.bloc
     root = self.root
 
-    if not (colon := prev_find(root._lines, *prev_bound(b0), b0_ln, b0_col, ':', True, comment=True, lcont=None)):  # must be there
-        return
+    if not (colon := prev_find(root._lines, *prev_bound(b0), b0_ln, b0_col, ':', True, comment=True, lcont=None)):  # only found if is on same logical line as first body child
+        return False
 
     if indent is None:
         indent = b0._get_block_indent()
@@ -990,6 +994,8 @@ def _normalize_block(self: fst.FST, field: str = 'body', *, indent: str | None =
     ln, col = colon
 
     self._put_src(['', indent], ln, col + 1, b0_ln, b0_col, False)
+
+    return True
 
 
 def _can_del_all(self: fst.FST, field: str, options: Mapping[str, Any]) -> bool: